requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.8.1
awscli==1.33.15
blinker==1.8.2
boto3==1.34.133
botocore==1.34.133
Brotli==1.1.0
cachetools==5.3.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
ConfigArgParse==1.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.8
distro==1.9.0
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.63.3
django-cors-headers==4.4.0
django-db-connection-pool==1.2.5
djangorestframework==3.15.2
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt==5.3.1
docutils==0.16
drf-yasg==1.21.7
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Login==0.6.3
gevent==24.2.1
geventhttpclient==2.3.1
google-auth==2.32.0
greenlet==3.0.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.2.0
itypes==1.2.0
Jinja2==3.1.4
jmespath==1.0.1
locust==2.31.2
MarkupSafe==2.1.5
msgpack==1.0.8
mysqlclient==2.2.4
openai==1.40.0
openapi-codec==1.3.2
packaging==24.1
pluggy==1.5.0
psutil==6.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
pytest==8.2.2
pytest-django==4.8.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.1.0
requests==2.32.3
rsa==4.7.2
s3transfer==0.10.2
sentry-sdk==2.12.0
setuptools==69.5.1
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
sqlparams==6.0.1
sqlparse==0.5.0
tqdm==4.66.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.6
Werkzeug==3.0.3
wheel==0.43.0
zope.event==5.0
zope.interface==7.0.1
